Fix tanh

+++ core-entities/src/commonMain/kotlin/org/pointyware/ember/entities/activations/Tanh.kt
@@ -7,11 +7,11 @@ import kotlin.math.tanh
  * Calculates the hyperbolic tangent of the input value.
  */
 object Tanh: ScalarActivationFunction {
-    override fun scalarActivation(input: Double): Double {
+    override fun scalarActivation(input: Float): Float {
         return tanh(input)
     }
 
-    override fun scalarDerivative(input: Double): Double {
+    override fun scalarDerivative(input: Float): Float {
         return 1 - tanh(input).pow(2)
     }
 }
